// frontend/src/company.d.ts
export interface CompanyKeyMetrics {
  metric: {
    marketCapitalization: number;
    currentRatioTTM: number;
    roeTTM: number;
    assetTurnoverTTM: number;
    cashFlowPerShareTTM: number;
    bookValuePerShareAnnual: number;
    currentDividendYieldTTM: number;
    capexCagr5Y: number;
    peBasicExclExtraTTM: number;
  };
}